chore: separate formatting from serialization file
