+++ libs/coin-algorand/src/types.ts
@@ -2,6 +2,8 @@ import type {
   Account,
   AccountRaw,
   Operation,
+  OperationExtra,
+  OperationExtraRaw,
   OperationRaw,
   TransactionCommon,
   TransactionCommonRaw,
@@ -55,11 +57,20 @@ export type AlgorandOperationExtra = {
   memo?: string;
   assetId?: string;
 };
+export function isAlgorandOperationExtra(op: OperationExtra): op is AlgorandOperationExtra {
+  return "rewards" in op || "memo" in op || "assetId" in op;
+}
+
 export type AlgorandOperationExtraRaw = {
   rewards?: string;
   memo?: string;
   assetId?: string;
 };
+export function isAlgorandOperationExtraRaw(
+  op: OperationExtraRaw,
+): op is AlgorandOperationExtraRaw {
+  return "rewards" in op || "memo" in op || "assetId" in op;
+}
 
 export type AlgorandAccount = Account & {
   algorandResources: AlgorandResources;
