+++ libs/coin-polkadot/src/bridge/createTransaction.ts
@@ -1,0 +1,3 @@
+import { createTransaction } from "../logic/transaction";
+
+export default createTransaction;
