+++ libs/ledger-live-common/src/account/formatters.ts
@@ -10,7 +10,7 @@ import { getOperationAmountNumberWithInternals } from "../operation";
 import { formatCurrencyUnit } from "../currencies";
 import { getOperationAmountNumber } from "../operation";
 import { getTagDerivationMode } from "@ledgerhq/coin-framework/derivation";
-import byFamily from "../generated/account";
+import byFamily from "../generated/formatters";
 import { nftsByCollections } from "@ledgerhq/live-nft";
 import type { Unit } from "@ledgerhq/types-cryptoassets";
 import type { Account, Operation, ProtoNFT } from "@ledgerhq/types-live";
