+++ libs/ledger-live-common/src/bridge/impl.ts
@@ -25,28 +25,41 @@ export const getCurrencyBridge = (currency: CryptoCurrency): CurrencyBridge => {
     currencyName: currency.name,
   });
 };
+
 export const getAccountBridge = (
   account: AccountLike,
   parentAccount?: Account | null,
 ): AccountBridge<any> => {
   const mainAccount = getMainAccount(account, parentAccount);
   const { currency } = mainAccount;
-  const { family } = currency;
-  const { type } = decodeAccountId(mainAccount.id);
   const supportedError = checkAccountSupported(mainAccount);
 
   if (supportedError) {
     throw supportedError;
   }
 
-  if (type === "mock") {
-    const mockBridge = mockBridges[currency.family];
-    if (mockBridge) return mockBridge.accountBridge;
+  try {
+    return getAccountBridgeByFamily(currency.family, mainAccount.id);
+  } catch {
+    throw new CurrencyNotSupported("currency not supported " + currency.id, {
+      currencyName: currency.name,
+    });
+  }
+};
+
+export function getAccountBridgeByFamily(family: string, accountId?: string): AccountBridge<any> {
+  if (accountId) {
+    const { type } = decodeAccountId(accountId);
+
+    if (type === "mock") {
+      const mockBridge = mockBridges[family];
+      if (mockBridge) return mockBridge.accountBridge;
+    }
   }
 
   const jsBridge = jsBridges[family];
-  if (jsBridge) return jsBridge.accountBridge;
-  throw new CurrencyNotSupported("currency not supported " + currency.id, {
-    currencyName: mainAccount.currency.name,
-  });
-};
+  if (!jsBridge) {
+    throw new CurrencyNotSupported("currency not supported " + family);
+  }
+  return jsBridge.accountBridge;
+}
