+++ libs/ledger-live-common/src/families/celo/bridge/js.ts
@@ -10,7 +10,12 @@ import broadcast from "../js-broadcast";
 import estimateMaxSpendable from "../js-estimateMaxSpendable";
 import prepareTransaction from "../js-prepareTransaction";
 import createTransaction from "../js-createTransaction";
-import { assignFromAccountRaw, assignToAccountRaw } from "../serialization";
+import {
+  assignFromAccountRaw,
+  assignToAccountRaw,
+  fromOperationExtraRaw,
+  toOperationExtraRaw,
+} from "../serialization";
 
 const receive = makeAccountBridgeReceive();
 
@@ -32,6 +37,8 @@ const accountBridge: AccountBridge<Transaction> = {
   broadcast,
   assignFromAccountRaw,
   assignToAccountRaw,
+  fromOperationExtraRaw,
+  toOperationExtraRaw,
 };
 export default {
   currencyBridge,
