+++ libs/ledger-live-common/src/families/celo/types.ts
@@ -2,6 +2,8 @@ import type {
   Account,
   AccountRaw,
   Operation,
+  OperationExtra,
+  OperationExtraRaw,
   OperationRaw,
   TransactionCommon,
   TransactionCommonRaw,
@@ -141,7 +143,13 @@ export type CeloOperationExtra = {
   celoOperationValue: BigNumber;
   celoSourceValidator?: string;
 };
+export function isCeloOperationExtra(op: OperationExtra): op is CeloOperationExtra {
+  return "celoOperationValue" in op;
+}
 export type CeloOperationExtraRaw = {
   celoOperationValue: string;
   celoSourceValidator?: string;
 };
+export function isCeloOperationExtraRaw(op: OperationExtraRaw): op is CeloOperationExtraRaw {
+  return "celoOperationValue" in op;
+}
