+++ libs/ledger-live-common/src/families/cosmos/serialization.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from "bignumber.js";
-import type { CosmosResourcesRaw, CosmosResources, CosmosAccountRaw, CosmosAccount } from "./types";
-import { Account, AccountRaw } from "@ledgerhq/types-live";
+import { type CosmosResourcesRaw, type CosmosResources, type CosmosAccountRaw, type CosmosAccount, isCosmosOperationExtraRaw, CosmosOperationExtra, CosmosOperationExtraRaw } from "./types";
+import { Account, AccountRaw, OperationExtra, OperationExtraRaw } from "@ledgerhq/types-live";
 
 export function toCosmosResourcesRaw(r: CosmosResources): CosmosResourcesRaw {
   const {
@@ -93,4 +93,75 @@ export function assignFromAccountRaw(accountRaw: AccountRaw, account: Account) {
   const cosmosResourcesRaw = (accountRaw as CosmosAccountRaw).cosmosResources;
   if (cosmosResourcesRaw)
     (account as CosmosAccount).cosmosResources = fromCosmosResourcesRaw(cosmosResourcesRaw);
+}
+
+export function fromOperationExtraRaw(extraRaw: OperationExtraRaw): OperationExtra {
+  if (!isCosmosOperationExtraRaw(extraRaw)) {
+    throw new Error();
+  }
+
+  const extra: CosmosOperationExtra = {};
+  if (extraRaw.validator) {
+    extra.validator = {
+      address: extraRaw.validator.address,
+      amount: new BigNumber(extraRaw.validator.amount),
+    };
+  }
+
+  if (extraRaw.validators && extraRaw.validators.length > 0) {
+    extra.validators = extraRaw.validators.map(validator => ({
+      address: validator.address,
+      amount: new BigNumber(validator.amount),
+    }));
+  }
+
+  if (extraRaw.sourceValidator) {
+    extra.sourceValidator = extraRaw.sourceValidator;
+  }
+
+  if (extraRaw.autoClaimedRewards) {
+    extra.autoClaimedRewards = extraRaw.autoClaimedRewards;
+  }
+
+  if (extraRaw.memo) {
+    extra.memo = extraRaw.memo;
+  }
+
+  return extra;
+}
+
+export function toOperationExtraRaw(extra: OperationExtra): OperationExtraRaw {
+  if (!isCosmosOperationExtraRaw(extra)) {
+    throw new Error();
+  }
+
+  const extraRaw: CosmosOperationExtraRaw = {};
+
+  if (extra.validator) {
+    extraRaw.validator = {
+      address: extra.validator.address,
+      amount: extra.validator.amount.toString(),
+    };
+  }
+
+  if (extra.validators && extra.validators.length > 0) {
+    extraRaw.validators = extra.validators.map(validator => ({
+      address: validator.address,
+      amount: validator.amount.toString(),
+    }));
+  }
+
+  if (extra.sourceValidator) {
+    extraRaw.sourceValidator = extra.sourceValidator;
+  }
+
+  if (extra.autoClaimedRewards) {
+    extraRaw.autoClaimedRewards = extra.autoClaimedRewards;
+  }
+
+  if (extra.memo) {
+    extraRaw.memo = extra.memo;
+  }
+
+  return extraRaw;
 }
