+++ libs/ledger-live-common/src/families/elrond/formatters.ts
@@ -3,12 +3,7 @@ import type { Operation } from "@ledgerhq/types-live";
 import { getAccountUnit } from "../../account";
 import { formatCurrencyUnit } from "../../currencies";
 import type { Unit } from "@ledgerhq/types-cryptoassets";
-import type {
-  ElrondAccount,
-  ElrondOperation,
-  ElrondOperationExtra,
-  ElrondOperationExtraRaw,
-} from "./types";
+import type { ElrondAccount, ElrondOperation } from "./types";
 import BigNumber from "bignumber.js";
 
 function formatAccountSpecifics(account: ElrondAccount): string {
@@ -75,25 +70,7 @@ function formatOperationSpecifics(op: Operation, unit: Unit | null | undefined):
     : "";
 }
 
-export function fromOperationExtraRaw(extraRaw: ElrondOperationExtraRaw) {
-  const extra: ElrondOperationExtra = {};
-  if (extraRaw.amount) {
-    extra.amount = new BigNumber(extraRaw.amount);
-  }
-  return extra;
-}
-
-export function toOperationExtraRaw(extra: ElrondOperationExtra) {
-  const extraRaw: ElrondOperationExtraRaw = {};
-  if (extra.amount) {
-    extraRaw.amount = extra.amount.toString();
-  }
-  return extraRaw;
-}
-
 export default {
   formatAccountSpecifics,
   formatOperationSpecifics,
-  fromOperationExtraRaw,
-  toOperationExtraRaw,
 };
